requirements.txt
numpy
matplotlib.pyplot
scipy
os